tests/wpt/web-platform-tests/css/css-toggle/parsing/toggle-visibility-invalid.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: parsing toggle-visibility with invalid values</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-visibility-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<body>
<script>

test_invalid_value('toggle-visibility', 'none');
test_invalid_value('toggle-visibility', 'toggle');
test_invalid_value('toggle-visibility', 't');
test_invalid_value('toggle-visibility', 'normal, toggle t');
test_invalid_value('toggle-visibility', 'toggle t, normal');
test_invalid_value('toggle-visibility', 'toggle a, toggle t');

</script>
